requirements.txt
hyperliquid-python-sdk>=0.1.0
pandas>=1.5.3
numpy>=1.24.0
requests>=2.28.0
eth-account>=0.8.0
websocket-client>=1.5.0
python-dotenv>=1.0.0
prettytable>=3.5.0